Cleaning up some casting warnings, including some legitimate 64-bit bugs. Not even gonna touch the JIT ones yet. The JIT overall is a bit of a mess. These cartridges aren't large. The entire program can be statically transcoded. Why even bother with a JIT? Additional overhead and complexity for no reason.

+++ desmume/src/frontend/windows/main.cpp
@@ -6719,7 +6719,7 @@ const char* MakeScriptPathAbsolute(const char* filename, const char* extraDirToC
 	return filename;
 }
 
-extern void RequestAbortLuaScript(int uid, const char* message);
+extern void RequestAbortLuaScript(uintptr_t uid, const char* message);
 
 const char* OpenLuaScript(const char* filename, const char* extraDirToCheck, bool makeSubservient)
 {
